pyproject.toml
[project]
name = "jp-stock-ml"
version = "0.1.0"
description = "日本株機械学習パイプライン - 成長銘柄発掘"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Data Collection & Processing
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "beautifulsoup4>=4.12.0",  # TDnetスクレイピング
    "lxml>=4.9.0",             # XMLパース
    "yfinance>=0.2.0",         # 海外指数
    "pytrends>=4.9.0",         # Google Trends
    
    # Database
    "sqlalchemy>=2.0.0",
    
    # Technical Analysis
    "ta>=0.10.2",
    
    # Machine Learning (Phase 2以降)
    "lightgbm>=4.0.0",
    "scikit-learn>=1.3.0",
    "optuna>=3.3.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
